feat(ratings): half-step ratings + tier-list page
Ratings now support 0.5 steps (Numeric(2,1)) with half-star input and a
clear affordance. Adds a "My Ratings" tier-list page that groups rated
books into themeable tiers (UR/SSR, S/A+, 夯到拉…), with the per-user
theme stored on the user. 夯到拉 is only offered in Chinese locales.

Backend:
- migration 039: rating SmallInteger -> Numeric(2,1); add users.tier_theme
- validate rating 0.5-5 in 0.5 steps; schemas/models -> float
- GET /books/all?has_rating filter + user_rating on each book
- PUT /auth/tier-theme; tier_theme exposed on /auth/me
- tests: tier-theme endpoint + has_rating param

Frontend:
- StarRating: half-star hover/click, click-to-clear + clear button
- tiers.ts: presets + rating->tier mapping (rating stays canonical)
- /ratings tier table + theme picker; nav entries (sidebar + profile)

+++ backend/app/schemas/reading.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 
 
 class RatingUpdate(BaseModel):
-    rating: int | None = None  # 1-5 or null
+    rating: float | None = None  # 0.5-5 in 0.5 steps, or null
 
 
 class FavoriteUpdate(BaseModel):
@@ -99,7 +99,7 @@ class ReadingGoalUpdate(BaseModel):
 
 
 class InteractionOut(BaseModel):
-    rating: int | None
+    rating: float | None
     is_favorite: bool
     reading_progress: dict | None
     reading_status: str | None
